+++ src/window_main/topNav.tsx
@@ -1,18 +1,39 @@
 import _ from "lodash";
-import mountReactComponent from "./mountReactComponent";
-import { createDiv } from "../shared/dom-fns";
 import React from "react";
 import ReactDOM from 'react-dom';
+import { queryElements as $$ } from "../shared/dom-fns";
+import { openTab } from "./tabControl";
+
+import {
+    MAIN_HOME ,
+    MAIN_DECKS,
+    MAIN_HISTORY,
+    MAIN_EVENTS,
+    MAIN_EXPLORE,
+    MAIN_ECONOMY,
+    MAIN_COLLECTION,
+    MAIN_CONSTRUCTED,
+    MAIN_LIMITED
+} from "../shared/constants";
 
 interface topNavItemProps {
-    id: string,
+    id: number,
     title: string
 }
 
 function TopNavItem(props:topNavItemProps) {
     const {id, title} = props;
+
+    const clickTab = React.useCallback((tabId: number) => (event: React.MouseEvent<HTMLDivElement>) => {
+        if (!event.currentTarget.classList.contains("item_selected")) {
+            $$(".top_nav_item").forEach(el => el.classList.remove("item_selected"));
+            event.currentTarget.classList.add("item_selected");
+            openTab(tabId);
+        }
+    }, [props.id]);
+
     return (
-        <div className={"top_nav_item it" + id}>
+        <div className={"top_nav_item it" + id} onClick={clickTab(id)}>
             {title !== "" ? (<span className={"top_nav_item_text"}>{title}</span>) : false}
             <div className={"top_nav_icon icon_" + id} title={_.camelCase(title)}></div>
         </div>
@@ -20,16 +41,16 @@ function TopNavItem(props:topNavItemProps) {
 }
 
 function TopNav() {
-    const homeTab = {id: "h", title:""};
-    const myDecksTab = {id: "0", title:"MY DECKS"};
-    const historyTab = {id: "1", title:"HISTORY"};
-    const eventsTab = {id: "2", title:"EVENTS"};
-    const exploreTab = {id: "3", title:"EXPLORE"};
-    const economyTab = {id: "4", title:"ECONOMY"};
-    const collectionTab = {id: "5", title:"COLLECTION"};
+    const homeTab = {id: MAIN_HOME, title:""};
+    const myDecksTab = {id: MAIN_DECKS, title:"MY DECKS"};
+    const historyTab = {id: MAIN_HISTORY, title:"HISTORY"};
+    const eventsTab = {id: MAIN_EVENTS, title:"EVENTS"};
+    const exploreTab = {id: MAIN_EXPLORE, title:"EXPLORE"};
+    const economyTab = {id: MAIN_ECONOMY, title:"ECONOMY"};
+    const collectionTab = {id: MAIN_COLLECTION, title:"COLLECTION"};
 
-    const contructedNav = {id: "7", title:""};
-    const limitedNav = {id: "8", title:""};
+    const contructedNav = {id: MAIN_CONSTRUCTED, title:""};
+    const limitedNav = {id: MAIN_LIMITED, title:""};
 
     return (
         <div className={"top_nav"}>
@@ -58,6 +79,5 @@ export default function createTopNav(parent: Element): boolean {
         <TopNav />,
         parent
     );
-
     return true;
 }
